packages/ars_analysis/pyproject.toml
[project]
name = "ars-analysis"
version = "2.0.0"
description = "ARS (Account Revenue Solution) v2 modular pipeline: 20 analytics modules, batch processing, Excel + PPTX output"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "pandas>=2.2,<3.0",
    "numpy>=1.26,<3.0",
    "matplotlib>=3.10,<4.0",
    "seaborn>=0.13",
    "openpyxl>=3.1.5",
    "python-pptx>=1.0",
    "python-dateutil>=2.8",
    "Pillow>=10.0",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.12",
    "loguru>=0.7",
    "typer[all]>=0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ars_analysis"]
